Board-Config: Enable Reporting of Current (mA) on LockScreen
Note---
With this Flag, "current_now" and "voltage_now" sysFS are used in place of "current_max" and "voltage_max". This allows for Showing Real-Time Current in mA on LockScreen while Charging.

I have Failed to Find the Original Author of this Commit. If you know, then please Mention in the Comments' Section below so that I can Update the Authorship on this Commit.